Add types to simple pages
Summary:
As discussed w/ @nlanam - updating page types to be of type React.FC. Replacing some JSX.Elements and adding to pages that didn't have it before.

Some pages have withStyles - but those are not cooperating with the type system. @nlanam suggests we switch to useStyles for those pages.

Test Plan: yarn dev

Reviewers: nlanam, michelle

Reviewed By: michelle

Subscribers: nlanam

Differential Revision: https://phab.corp.pixielabs.ai/D8329

GitOrigin-RevId: 1b0fb208cf4f6ff074156005d832d6e11cb23291