Fewer calls to now(): Part 1 (Prep)
Summary:
This is prepatory work. It pushes the calls to now() up higher in the call stack. For now there are no savings,
but the eventual goal of this work is to consolidate now() calls across trackers, by having a single shared expiry_timestamp per iteration.

Test Plan: Existing tests.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

JIRA Issues: PP-2647

Differential Revision: https://phab.corp.pixielabs.ai/D8343

GitOrigin-RevId: 57ec78692c8ad53aba587d1ccb50cde4e5205b9e