scripts/create_gke_cluster.sh: add optional argument -z for zone.
Summary:
scripts/create_gke_cluster.sh: add optional argument -z for zone.

We are adding an argument "-z" to specifiy the zone for the cluster. This is useful for the stirling perf eval flow because we want to run multiple perf evals in parallel (as a single experiement where we have a baseline and multiple perf experiments to evaluate). For isolation, sandboxing, and reproducibility, we spin-up and tear-down a new cluster for the individual experiments. Zone us-west1-a is oversubscribed by our team. Allowing the stirling perf eval flow to choose a different zone (it will use us-west1-b for now) enables the desired parallelism.

Test Plan: Use case by stirling perf eval.

Reviewers: #stirling, zasgar, vihang, michelle, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2555

Differential Revision: https://phab.corp.pixielabs.ai/D8422

GitOrigin-RevId: 67b329d88f30ccbc7a99ea9993faff5e6351e0ae