PP-2684 Don't include namespace in helm chart/extracted YAMLs
Summary:
now that we're updating vizier so that it can be deployed to any namespace, we should follow helm chart conventions a little better.
in general, namespaces are not created in the helm charts. it can be created using the --create-namespace flag on the helm command or should already exist.

this diff updates the generated template so that it no longer includes the namespace resource. if theyre deploying through helm, helm will handle the namespace for them. if they're deploying through --extract_yamls, we can probably assume that they know how to create their own namespace first.

if deploying through the CLI, we can still create the namespace for them if needed.

Note: This will break new Vizier releases for older CLIs, since the old CLIs will still want the namespace YAML! however, more breaking changes are coming in followup diffs, so it's probably not worth extra changes to make sure this doesn't break old CLIs.

Test Plan: ran `px deploy` with/without existing pl namespace.

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2684

Differential Revision: https://phab.corp.pixielabs.ai/D8431

GitOrigin-RevId: b170e1ed0e4a6a85ba7b081c1309a667225dfeed