PP-2684 Update vizier templates to allow deploying in any namespace
Summary:
We want viziers to be deployable to any namespace. this diff templatizes the vizier YAMLS so that the namespace is configurable.
For most of our template values, we use `.Values.templateField`. For namespace, we use `.Release.Namespace` so that it works with Helm chart's `--namespace` flag.
As a result, this will break existing CLIs, because the new `.Release` field will confuse the templater. However, we should prefer to follow standard conventions.

Test Plan:
tested that it works in all of the following cases:
1. updating 0.7.1 -> rc (default, old pl namespace)
2. updating rc -> rc (both pl namespace and new namespace)
3. extract yaml (new namespace)
4. helm chart (new namespace)
exciting screenshot:
{F154569}

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2684

Differential Revision: https://phab.corp.pixielabs.ai/D8433

GitOrigin-RevId: 6f46c975739f69506185fa8fa204aef9103c77f7