PP-2684 Update CLI delete/collectLogs to detect vizier namespace
Summary:
now that we allow vizier to run in any namespace, we should update the CLI delete/collectLogs command to detect the vizier namespace rather than relying on the user to pass it in as a flag.
it does this by searching for any pods with the vizier label. in the future when we have our CRD, this will be a little cleaner since we can just look for the CRD than some random pod.

Test Plan: deployed vizier to a non-pl namespace, ran both CLI delete and collectLogs.

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8436

GitOrigin-RevId: f0e08d6773dafdaa77086eb993fca6e323fc68d9