Allow pinned vizier/cli versions in artifact tracker
Summary:
if a vizier/cli version is specified in the artifact tracker deployment, it assumes that the version exists without having to check the db.
uesrs who are deploying a fresh pixie cloud will not need to run load_dev_db.

Test Plan: deployed to a version to staging and tested CLI updates/vizier deploys

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8505

GitOrigin-RevId: 983439503e3b87842c4d5a2c17552a6c92e26d97