CGroupMetadataReader: try various cgroup paths for robustness
Summary:
A user reported issues deploying the PEM. It turned out their Linux distro--WSL--didn't have /sys/fs/cgroup/cpu,cpuacct mounted.

As a workaround, also search /sys/fs/cgroup/cpu. We are just looking for PIDs, so it should work as well.

Test Plan: Not tested yet. Need to find a WSL system to try this on. In the meanwhile, it shouldn't break functionality on existing systems.

Reviewers: michelle, zasgar, yzhao

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8519

GitOrigin-RevId: 721a0a29dbfae9fc917d9e99f27e8a5bec7d7a2a