PC-883: Re-enable CORS checks for cli redirect URI by passing token in header
Summary:
CORS checks were not working as expected. Browser CORS is enabled in a complex set of circumstances and we did not hit those circumstances with our GET request. We found that we could enable CORS by adding a header to the GET request.
It seems like we had plans to pass the data in the `token` header, so we are now passing the token there. In a follow up diff, we will read the data from the header in the CLI server. Then we will disable the old method of passing the accessToken through the URL.

D8524 enables the new CLI path for this.

Test Plan: Tested against sketchy URLs and we stopped sending over the request. We still work with localhost redirect URIs which is intended

Reviewers: zasgar, vihang

Reviewed By: zasgar

JIRA Issues: PC-883

Differential Revision: https://phab.corp.pixielabs.ai/D8523

GitOrigin-RevId: 6b0c5ddd0450f02132d1e83c72932db0e7efa803