[PC-693] Resize quantile charts in data tables to always fit their containing cells
Summary: Does what it says on the tin, and fixes nearby lint warnings (arc lint enforces this).

Test Plan: Look at a data table with a column that shows quantiles. For example, `px/namespace` in the Service List table. Resize the column to be tiny, huge, and in between. It should always show both the quantile chart and the actual data value, rather than cutting either off.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-693

Differential Revision: https://phab.corp.pixielabs.ai/D8532

GitOrigin-RevId: f0449ef33c05d619d3234c8975650d7debff971d