PC-894 Fix ListClusters call when user has no deployed viziers
Summary:
the UI's call to ListClusters was erroring when the user has no deployed clusters with `empty slice passed to 'in' query`.
this was because we were trying to pass an empty array to the sqlx 'in' query, which is not allowed. updated the code to shortcircuit if the array is empty.

Test Plan: tried on staging with UI changes from D8562

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang

JIRA Issues: PC-889

Differential Revision: https://phab.corp.pixielabs.ai/D8563

GitOrigin-RevId: 90cc2c448ffb3e3677489f320037362ba0436c50