Move dev_cloud_namespace to the root command
Summary:
The root command overrides `dev_cloud_namespace` so we should define
the flag at the root level and let all `px` commands inherit it. Also check the
flag being set instead of it's value in root PreRun when overriding it.

(The existence of the same flag in load-cluster-secrets was causing some issues,
load-cluster-secrets is gone now and hence the issue is gone too, but this is a
better way to handle this flag anyway and will prevent future issues.)

Test Plan:
Logged the flag value in `deploy.go`. Observed that setting `PL_TESTING_ENV`
caused the namespace to be set. However setting the flag explicitly also worked
and preferentially used the user set value.

Reviewers: zasgar, michelle, nserrino

Reviewed By: zasgar

JIRA Issues: PC-904

Differential Revision: https://phab.corp.pixielabs.ai/D8572

GitOrigin-RevId: ba2f8bd6071d33b13518e358ecc7e866d847f509