perf_profiler: Symbolizer class to own an instance of ebpf::BPFStackTable & improve tests.
Summary: In the stirling perf-profiler, refactor the symbolizer class such that it creates and owns an instance of ebpf::BPFStackTable, i.e. for the sole purpose of gaining access to the BCC symbolization API. Previously, we inadvertantly and inefficiently used two instances of the same because of our double buffering technique. This change also improves the test cases for the symbolizer and for the perf profile connector.

Test Plan: Existing (and improved) tests.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2688

Differential Revision: https://phab.corp.pixielabs.ai/D8584

GitOrigin-RevId: 44a53af140d60edcbb2e40b379c167cfea1dd8c0