Don't warnError in bazelCICmd
Summary:
`bazelCICmd` can be run in k8s or docker. For the former, we retry
if we catch k8s cluster downscale errors. So the command failure isn't treated
as an error if it's of type cluster downscale. Since `WithSourceCodeK8s`
already handles the `warnError` wrapping, we don't need to.
For the latter, we add a `warnError` wrapper in `WithSourceCodeAndTargetsDocker`
instead.

Test Plan: jenkins

Reviewers: zasgar, michelle

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8652

GitOrigin-RevId: 42f28ef1fcce5ecd9a6962121a1f4988c1d99f25