PC-917: Don't render sidebar components if the ClusterContext is undefined
Summary: ClusterContext will not be defined for some sidebar owners. That's because we want to rely on the URL as the source of truth for the cluster, but the URLs in /credits and /admin do not contain this info, nor should they contain this info. Will have to update in a later diff to rethink what goes in the sidebar components for these and other pages.

Test Plan: Tested on yarn dev with a user that has 0 clusters. Tested on the simplified routing approach and both work.

Reviewers: vihang, zasgar, michelle

Reviewed By: zasgar

JIRA Issues: PC-917

Differential Revision: https://phab.corp.pixielabs.ai/D8685

GitOrigin-RevId: 2e5e46a2b7944b898895a4db1dfe9f7a8178f05a