PP-2751: px/service should work for services that come in pairs
Summary: Previously, px/service checked for service string equality (e.g. df.service == 'ns/sv1'). This caused data to now show up for records that are part of multiple services. This change fixes that problem by switching the equality checks to px.has_service_name, which allows for checking if a service is in the list of services.

Test Plan: ran it with a service that comes in a pair.

Reviewers: michelle, philkuz, zasgar, vihang

Reviewed By: michelle

JIRA Issues: PP-2751

Differential Revision: https://phab.corp.pixielabs.ai/D8687

GitOrigin-RevId: 222576192de80021d8e390e3a956074c5edfe292