Add gql endpoint to get clusterInfo by name
Summary:
In many places in the UI, we reference the cluster by name (for example, URL).
however, the backend only provides a way to fetch a single ClusterInfo by ID. therefore, consumers who need clusterInfo but only have a clusterName currently fetch the entire list of clusters, and filter through the list to get the correct clusterInfo.
Instead, we can clean this logic up by doing this filtering in the backend.

Test Plan: unit test

Reviewers: vihang, philkuz

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8691

GitOrigin-RevId: c2c80b16292a476adca197c4ccbcd80bf66d1f54