PC-940 Actually fix topbar when showing deployInstructions
Summary: D8689 was supposed to fix the topbar in the deployinstructions... It did fix it in the deployInstructions, but didn't work when in an org with clusters. this is because we shouldn't actually make topbar absolute, because other content will go underneath the topbar.

Test Plan: checked org with and without clusters

Reviewers: vihang, philkuz, zasgar

Reviewed By: vihang

JIRA Issues: PC-940

Differential Revision: https://phab.corp.pixielabs.ai/D8692

GitOrigin-RevId: 797caafe2e993526295d9e46913f89355994e68d