PP-2737: Vizier pods (besides bootstrap) should wait for cloud connector to assign cluster ID before starting
Summary: We had a problem where users would be confused when deploying via helm, becuase many Vizier pods would show up as CreateConfigError. This was due to the fact that a subset of vizier pods (non-bootstrap pods) are configured to require PL_CLUSTER_ID, which is assigned in the bootstrap phase by cloud-connector. This diff adds a /readyz to cloud-connector which checks whether or not PL_CLUSTER_ID has been assigned. /readyz is then used by the non-bootstrap pods to determine whether or not the pod should start up yet (having received PL_CLUSTER_ID by that point).

Test Plan: ran it in skaffold various times

Reviewers: michelle, zasgar, vihang

Reviewed By: michelle

JIRA Issues: PP-2737

Differential Revision: https://phab.corp.pixielabs.ai/D8712

GitOrigin-RevId: 9cef9278b2dbf919f6bcd02b1aa239e13ca2adfa