Warn the user that px delete without clobber will delete the entire namespace. Add y/n verification for delete to match deploy.
Summary: TSIA. If we do px delete, we will delete the whole namespace and the user should give y/n confirmation. Also we should tell them that px delete is going to do that, in case they deployed px to a namespace with other things in it, like a monitoring namespace.

Test Plan: ran it

Reviewers: michelle, zasgar, vihang

Reviewed By: michelle, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8718

GitOrigin-RevId: 7c7436ed73996d582edff506d874c942a3beddc5