[PC-956] Fix sidebar toggle button
Summary: Minor oversight since the function it calls serves double duty

Test Plan: Open the live tour with the sidebar already closed; then try with it already open. Also try opening and closing it from the top left hamburger button. It should close for the tour if open, remember what state it was in when the tour closes, and the hamburger button should switch open/closed as long as said button is available (the tour hides/blocks it).

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-956

Differential Revision: https://phab.corp.pixielabs.ai/D8732

GitOrigin-RevId: 350fa43d52be5fa451d0d9012b1380e43a395848