[PC-962] Drop focus on breadcrumbs when closed
Summary: Not doing this was causing the arrows keys and enter to retain focus on a hidden breadcrumb, which in the case of the script dropdown could unexpectedly re-run the current script or change it entirely. This change also fixes a bunch of lint warnings.

Test Plan: Open a breadcrumb (like that of the script), select something, then use the arrow keys and/or enter. Doing so should no longer behave as if that breadcrumb was still open and in-focus.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-962

Differential Revision: https://phab.corp.pixielabs.ai/D8749

GitOrigin-RevId: 8034cea8bde6d46a9f3a1a66dab4f4fb438e8fff