PC-966: Adding useQuery into the pixie-api-react package as a proxy for apollo
Summary: useQuery couldn't be imported in the main ui package becuase it created a new Apollo global state that was not compatible with the existing one used in the `pixie-api-react` package. This avoids that problem by just exposing useQuery as a member of our react package. Eventually we will swap this over to the normal apollo import, but that's when we completely finish up PC-966

Test Plan: Tested in a different diff by calling useQuery. Will follow up with that later.

Reviewers: michelle, vihang, nlanam, #third_party_approvers

Reviewed By: michelle, vihang, #third_party_approvers

JIRA Issues: PC-966

Differential Revision: https://phab.corp.pixielabs.ai/D8763

GitOrigin-RevId: f12c249eb7afe0f0e9fc057beca468d5f6d6ecdf