[PC-966] Remove `useIsAuthenticated` hook.
Summary:
Inlines it like we've been doing elsewhere.

Simply moved it to `app.tsx`, which is its only consumer at this time, and simplified it a bit. Added the `makeCancellable` method for Promises that was originally in D8515 (later reverted for other reasons) to make this easier.

Test Plan: Try logging in and out a few times. No behavioral change should be noticeable.

Reviewers: michelle, vihang, philkuz

Reviewed By: vihang

JIRA Issues: PC-966

Differential Revision: https://phab.corp.pixielabs.ai/D8837

GitOrigin-RevId: f725f90023840775505df93a47997dbf26588505