PC-986 Send id_token to all accessToken consumers from the UI
Summary:
Now we send over an id_token to all accessToken consumers from the UI. Not all auth providers currently create an idtoken and therefore not all will send it onwards. Shouldn't break anything for the moment.

you might notice that there's no solution for the manual copy to CLI. That's something I'll think more about in PC-988. IDToken is not consumed anywehre in the API at the moment so it shouldn't be a problem.

Test Plan: Made sure the network panel showed idToken as an argument in the UI

Reviewers: michelle, vjain

Reviewed By: michelle

JIRA Issues: PC-986

Differential Revision: https://phab.corp.pixielabs.ai/D8872

GitOrigin-RevId: 34f938b56431a8ad5359d365c0b815585c0f39d8