PC-985: Replace Auth0 javascript library with generic OIDC library and connect to Auth0
Summary:
Moving to a generic OIDC model. Here we convert the Auth0 model over to a more generic provider that can support any OIDC. We leverage the OIDC client library that handles lots of the logic for us, and we write up the implementation for the existing oauth-provider model. This whole process was rather simple.

Future developers of PIxie can choose to append more OIDC providers to their list of offerings. For Auth0 based OIDC, all they would need to do is call Auth0 with a new connection parameter. If they want to use a separate provider, they would just have to instantiate the OIDC definition for that specific provider.

Test Plan: Tested on yarn dev on a dev cluster and was able to sign up / login.

Reviewers: michelle, #third_party_approvers

Reviewed By: michelle, #third_party_approvers

JIRA Issues: PC-985

Differential Revision: https://phab.corp.pixielabs.ai/D8824

GitOrigin-RevId: 9e45c0959ae65ec178384e10ba00f9c9cab6ba74