PP-2445: Remove now-unnecessary px.head calls in scripts
Summary: Before PP-2445, we had to put in px.head calls in some scripts in order to ensure that Kelvin/PEM-only UDFs were scheduled on the correct Carnot instance. Now, this is unnecessary, because the compiler should be smarter about figuring it out.

Test Plan: ran each script in the UI

Reviewers: htroisi, philkuz, zasgar, michelle

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8881

GitOrigin-RevId: e69727728cfc772948fc4b1f244e3eb929646c30