PC-996: Migration that adds identity_provider column to users and defaults to google-oauth2
Summary: Adding two migrations 1 for identity_provider column and another to default to google-oauth2. A bit shaky on the second one, but not sure what a better option would be for us. The other option is to go and manually add this change in, but that also feels like a half solution.

Test Plan: Tested on skaffold and saw the migration work. Updated tests in datastore to make sure they work.

Reviewers: michelle, vjain, nserrino, vihang

Reviewed By: michelle

JIRA Issues: PC-996

Differential Revision: https://phab.corp.pixielabs.ai/D8904

GitOrigin-RevId: fdb008e93b7c3a11eca598ee8373f5cdaac72b4b