PC-1022: Fix script reference links that aren't entities from throwing a 404
Summary: After the massive refactor, we failed to account for links in our testing -> script_reference() for non entity references. The scriptEntityURL code was append a `script` field to the path which was not necessary as we store the script as an argument. We removed the `script` path hanlder in our massive refactor and now we have this.

Test Plan: Updated jest with new format. Also yarn dev, script references that previously 404'd now 200.

Reviewers: nlanam, michelle

Reviewed By: nlanam, michelle

JIRA Issues: PC-1022

Differential Revision: https://phab.corp.pixielabs.ai/D8920

GitOrigin-RevId: 472fbb208fa25358a0e66d24ae3018b7d4fa712a