Allow theme to be specified in query params
Summary:
we are adding the ability to embed Pixie in iframes. in order to better fit the theme of the site embedding Pixie, we are allowing users to specify the theme in the query params.
in the future, we should add support to toggle the theme based on a switch and user-settings. for now, the query param will suffice.

theme colors for the light theme are from: https://www.figma.com/file/WkM4NrbvWOZloJauYy54YP/Color-Palette-2.0?node-id=2%3A68
We should play around with the flamegraph colors to make those more presentable.

Test Plan:
screenshot:{F160403}
{F160419}

{F160420}

Reviewers: nlanam, vihang, zasgar

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D9010

GitOrigin-RevId: a86aa316cd8f0a2c7e11a6024c015b19becff986