Add widget selection support to UI
Summary: We want to be able to support embedding Live Views and Widgets via link. This diff adds an optional "widget" param to Live View URLs, which will support the case of embedding a particular widget via link. This is for the cases where you would rather show a single widget rather than the whole grid.

Test Plan: ran it in mobile, selected and unselected widget cases, updated unit tests.

Reviewers: zasgar, michelle, vihang, nlanam

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9016

GitOrigin-RevId: 87f0693e2468593796af30b9e3a7072febb0fcbe