PC-1051 Live view with no clusters should not throw an error
Summary:
Currently, we are throwing a snackbar error in the liveview if the user has no clusters.
this is because:
1. clusterName in live-routing is stringified. undefined clusterName becomes "undefined"
2. live.tsx makes a gql to fetch clusterInfo for "undefined". gql request returns an error since "undefined" does not exist.
3. since there is an error, we show the snackbar notification.

unfortunately, since the gql query is a react hook, we can't conditionally make the call. however, it also doesnt seem right to update the gql request to not return an error, because it /should/ throw an error if you're fetching clusterInfo for a non-existent cluster.

instead, if the clusterName is empty, we should know to expect an error and not show the snackbar. ideally we should also check the error message, but the error message from GQL is just a string rather than a nice error code.

Test Plan: yarn dev

Reviewers: vihang, nlanam, philkuz

Reviewed By: vihang

JIRA Issues: PC-1051

Differential Revision: https://phab.corp.pixielabs.ai/D9022

GitOrigin-RevId: 4c4580de4daa77eedb8b54f51da4cd5407b6c0ce