PC-1052 Make manual CLI auth token fit in width of auth box
Summary:
the auth token is inside a code renderer which overflows the size of the authbox.
since this token can get very long, stretching the size of the authbox to fit the token won't always work. instead, we can make the box scroll.
also updated the scrollbar styles so that it better matches the rest of our scrollbars.

Test Plan: yarn dev

Reviewers: nlanam, vihang

Reviewed By: nlanam

JIRA Issues: PC-1052

Differential Revision: https://phab.corp.pixielabs.ai/D9026

GitOrigin-RevId: 3e95b2b6fa4a1b145031172f4edb047fd6a76e8a