PP-2827 Don't show tour overlays for support accounts
Summary:
we currently always show a tour overlay for support accounts, because we try to fetch the user setting which doesnt exist.
we then try to write a user setting to record that the user has seen the tour. however, support accounts arent real users, so the query fails.

Test Plan: yarn dev

Reviewers: vihang, nlanam, nserrino

Reviewed By: nlanam

JIRA Issues: PP-2827

Differential Revision: https://phab.corp.pixielabs.ai/D9028

GitOrigin-RevId: dfaea121ea8fd849f64afb013055f91780886c81