PC-1054 Fix licenses in hosted Pixie
Summary:
I broke the licenses when making changes to deploy OSS pixie every OSS build. This should check the configurables/private path is a valid directory, rather than a valid file. This was always returning false, so wouldn't move the licenses to the correct path.
However, the licenses in the OSS build will still be broken. this is because the licenses build requires private creds. This is filed here: https://pixie-labs.atlassian.net/browse/PC-1055

Test Plan: should test with a deploy to staging

Reviewers: vihang, zasgar

Reviewed By: vihang

JIRA Issues: PC-1054

Differential Revision: https://phab.corp.pixielabs.ai/D9032

GitOrigin-RevId: d98d97abcfda91c2b4076d96ca6177e18841fdfc