Fetch segment info for the CLI from pixie cloud.
Summary:
This helps us differentiate between dev/staging/prod and also handles the OSS pixie deployment better.
This moves the viper flag parsing so that it happens in init and the cloud_addr is accessible by the time analytics client is created.

After this we will use cloud to check for analytics opt outs and use that in the CLI.

Test Plan: Ran cli with some logs around the analytics keys

Reviewers: zasgar, michelle

Reviewed By: michelle

JIRA Issues: PP-2626

Differential Revision: https://phab.corp.pixielabs.ai/D9071

GitOrigin-RevId: 22ea9c762dd8dfb9dde213a438af2d0f70fbfb9c