Add functionality to /embed to disable the time picker
Summary:
For some cases with /embed, we want the time to be fixed. (In the non-widget case that allows breadcrumbes).
In this scenario, we will omit the start_time breadcrumb from the breadcrumbs. This is controlled by a query param, and is not the default behavior.

fix test

Test Plan: ran various cases in UI, tests

Reviewers: zasgar, philkuz, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9090

GitOrigin-RevId: 5795598116bfedaa2e87b756135d72ec25510487