PP-2849 Update request graph to not show arrows pointing from entity with no pod/svc name
Summary: our current request graph can be confusing. it initially starts with the service view. in some cases, this shows nodes with no connections to each other. switching to the pod view almost always shows pods with connections to each other, which can be confusing. however, the pod that shows connections to other pods is actually a entity with no pod/service name (basically its some external entity we couldnt resolve). we should filter out this node, because it is confusing and not helpful

Test Plan: yarn dev

Reviewers: nserrino, zasgar, vihang

Reviewed By: vihang

JIRA Issues: PP-2849

Differential Revision: https://phab.corp.pixielabs.ai/D9128

GitOrigin-RevId: b7a7b2f119b434097dfccb36dab2dcd01777f557